+++ Main/src/SCP.cpp
@@ -233,7 +233,7 @@ formatError: ::SetLastError(ERROR_BAD_FORMAT);
 							pit->FlushSectorBuffers(); // convert all modifications into flux transitions
 							if (const DWORD trackLength=TrackToStream( *pit, CMemFile(buffer,SCP_BUFFER_CAPACITY), cylFile, head, bStreamAdjusted )){
 								// conversion of the Track to SuperCardPro stream succeeded?
-								const LONG capacity=((TTrackDataHeader *)buffer.get())->GetFullTrackCapacityInBytes( mp, header );
+								const DWORD capacity=((TTrackDataHeader *)buffer.get())->GetFullTrackCapacityInBytes( mp, header );
 								if (GetCurrentDiskFreeSpace()<capacity)
 									return ERROR_DISK_FULL;
 								// . success thanks to adjustments to overcome SuperCard Pro limitations?
